Refactor: Extract language-aware SendInVoiceRange per review
Move the per-listener message scrambling logic into a dedicated
SendInVoiceRangeWithLanguage method in a .SS220.cs sidecar, following
the reviewer's guidance to avoid inline loops and parallel the upstream
SendInVoiceRange structure. The method handles range checking, hideChat
propagation, and replay recording for each listener's variant of the
message.
